obsolete/back_end/requirements.txt
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
fhirclient==3.2.0
flasgger==0.9.5
Flask==1.1.2
Flask-Cors==3.0.9
Flask-PyMongo==2.3.0
idna==2.10
importlib-metadata==3.1.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mistune==0.8.4
pymongo==3.11.1
pyrsistent==0.17.3
python-dotenv==0.15.0
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
wincertstore==0.2
zipp==3.4.0
